backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
pydantic==2.5.0
python-multipart==0.0.6
websockets==12.0
openai==1.3.7
python-dotenv==1.0.0
geoalchemy2==0.14.2
httpx==0.25.2
tweepy==4.14.0
pandas==2.1.4
numpy==1.25.2
pytest==7.4.3
pytest-asyncio==0.21.1